Avoid including ASoA.h in ROOT dictionary generation
