+++ PWGEM/Dilepton/DataModel/dileptonTables.h
@@ -9,6 +9,8 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
+#include "EvSelFlags.h"
+
 #include "Common/CCDB/EventSelectionParams.h"
 #include "Common/Core/RecoDecay.h"
 #include "Common/DataModel/Centrality.h"
@@ -66,28 +68,6 @@ namespace o2::aod
 
 namespace emevsel
 {
-// Event selection criteria. See O2Physics/Common/CCDB/EventSelectionParams.h
-enum EventSelectionFlags {
-  kIsTriggerTVX = 0,          // FT0 vertex (acceptable FT0C-FT0A time difference) at trigger level
-  kNoITSROFrameBorder,        // bunch crossing is far from ITS RO Frame border
-  kNoTimeFrameBorder,         // bunch crossing is far from Time Frame borders
-  kNoSameBunchPileup,         // reject collisions in case of pileup with another collision in the same foundBC
-  kIsGoodZvtxFT0vsPV,         // small difference between z-vertex from PV and from FT0
-  kIsVertexITSTPC,            // at least one ITS-TPC track (reject vertices built from ITS-only tracks)
-  kIsVertexTOFmatched,        // at least one of vertex contributors is matched to TOF
-  kIsVertexTRDmatched,        // at least one of vertex contributors is matched to TRD
-  kNoCollInTimeRangeNarrow,   // no other collisions in specified time range (narrower than Strict)
-  kNoCollInTimeRangeStrict,   // no other collisions in specified time range
-  kNoCollInTimeRangeStandard, // no other collisions in specified time range with per-collision multiplicity above threshold
-  kNoCollInRofStrict,         // no other collisions in this Readout Frame
-  kNoCollInRofStandard,       // no other collisions in this Readout Frame with per-collision multiplicity above threshold
-  kNoHighMultCollInPrevRof,   // veto an event if FT0C amplitude in previous ITS ROF is above threshold
-  kIsGoodITSLayer3,           // number of inactive chips on ITS layer 3 is below maximum allowed value
-  kIsGoodITSLayer0123,        // numbers of inactive chips on ITS layers 0-3 are below maximum allowed values
-  kIsGoodITSLayersAll,        // numbers of inactive chips on all ITS layers are below maximum allowed values
-  kNsel                       // counter
-};
-
 DECLARE_SOA_BITMAP_COLUMN(Selection, selection, 32); //! Bitmask of selection flags
 DECLARE_SOA_DYNAMIC_COLUMN(Sel8, sel8, [](uint32_t selection_bit) -> bool { return (selection_bit & BIT(o2::aod::emevsel::kIsTriggerTVX)) && (selection_bit & BIT(o2::aod::emevsel::kNoTimeFrameBorder)) && (selection_bit & BIT(o2::aod::emevsel::kNoITSROFrameBorder)); });
 
@@ -149,7 +129,6 @@ uint32_t reduceSelectionBit(TBC const& bc)
   }
   return bitMap;
 }
-
 } // namespace emevsel
 
 namespace emevent
